requirements.txt
# Web Framework
fastapi==0.121.2
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.6.1

# LLM & AI
google-generativeai==0.8.3
sentence-transformers==3.3.1

# Vector Database
qdrant-client==1.12.1

# Document Processing
pymupdf==1.24.14
python-docx==1.1.2
#Pillow==10.4.0

# Utilities
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2